+++ bookmakerratings/src/config.ts
@@ -1,0 +1,11 @@
+import { Requester } from '@chainlink/external-adapter'
+import { Config } from '@chainlink/types'
+
+export const DEFAULT_ENDPOINT = 'game-odds'
+export const DEFAULT_BASE_URL = 'http://api.odds.ru/sfp/v1/en/'
+
+export const makeConfig = (prefix?: string): Config => {
+  const config = Requester.getDefaultConfig(prefix)
+  config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
+  return config
+}