+++ cryptomkt/src/config.ts
@@ -1,0 +1,11 @@
+import { Requester } from '@chainlink/external-adapter'
+import { Config } from '@chainlink/types'
+
+export const DEFAULT_ENDPOINT = 'ticker'
+export const DEFAULT_BASE_URL = 'https://api.cryptomkt.com/v1/'
+
+export const makeConfig = (prefix?: string): Config => {
+  const config = Requester.getDefaultConfig(prefix)
+  config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
+  return config
+}