+++ watchsignals/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as avgPrice from './avgPrice'